Séparations de classes en différents fichiers
